last fix

+++ app/models/entities/Session.go
@@ -8,8 +8,8 @@ import (
 // Session структура сессии пользователя
 type Session struct {
 	Token      string
-	UserID     int
-	EmployeeID int
+	UserID     int64
+	EmployeeID int64
 	Connection *sql.DB
 	Login      string
 	Created    time.Time
