+++ app/models/entities/User.go
@@ -2,8 +2,8 @@ package entities
 
 // User struct of user
 type User struct {
-	UDI      int
-	EID      int
+	UDI      int64
+	EID      int64
 	Login    string
 	Password string
 	Fullname string
